Analytic Query force refresh
added a "force_search" option to the TSeriesQuery.select() method. This will force search parameters on the system instead of looking for them in the saved cache. This is needed when a parameter that is similarly named to the intended one is already in the local cache but the one we want is not.